Rename functional test files
