+++ tests/test_appname_cmd.py
@@ -1,5 +1,5 @@
-from application_client.boilerplate_command_sender import KaspaCommandSender
-from application_client.boilerplate_response_unpacker import unpack_get_app_name_response
+from application_client.kaspa_command_sender import KaspaCommandSender
+from application_client.kaspa_response_unpacker import unpack_get_app_name_response
 
 
 # In this test we check that the GET_APP_NAME replies the application name
