+++ tests/test_name_version.py
@@ -1,5 +1,5 @@
-from application_client.boilerplate_command_sender import KaspaCommandSender
-from application_client.boilerplate_response_unpacker import unpack_get_app_and_version_response
+from application_client.kaspa_command_sender import KaspaCommandSender
+from application_client.kaspa_response_unpacker import unpack_get_app_and_version_response
 
 
 # Test a specific APDU asking BOLOS (and not the app) the name and version of the current app
