+++ tests/test_version_cmd.py
@@ -1,5 +1,5 @@
-from application_client.boilerplate_command_sender import KaspaCommandSender
-from application_client.boilerplate_response_unpacker import unpack_get_version_response
+from application_client.kaspa_command_sender import KaspaCommandSender
+from application_client.kaspa_response_unpacker import unpack_get_version_response
 
 # Taken from the Makefile, to update every time the Makefile version is bumped
 MAJOR = 0
